fix(renderer): round each squared radius at its assignment
Addresses the review on #91. The span-geometry conversions added in 027a2f1
rounded only `dy*dy`, which does not block anything: `radiusSquared` is itself a
product, so arm64 fused it straight into the subtraction as FNMSUBD and the
codegen still showed a fused op at every one of those lines. A span boundary
could therefore land differently than on amd64, and the incremental dirty mask
could include or omit a pixel that the full renderer did not.

Each squared radius is now rounded where it is assigned, in both the float64 and
float32 geometry paths and in the full-render scanline walker. No fused
multiply-add is left on any path that decides a pixel.

Four sites are left fused on purpose, because they choose search behaviour
rather than output: the circle-area budget heuristics in polish_dirty_cost.go
and incremental_cost.go, and the seed separation test and correctiveChannel in
batch_audit.go. docs/known-limitations.md records that, and that it means fits
are not bit-reproducible across architectures even though renders now are.

`r2` is renamed to `radiusSquared` in renderCircleScanlineRowsTracked because
adding the comment moved its declaration and varnamelen then reported it.

AGENTS.md said the ARM64 rows skip internal/fit/renderer and that
check-cross-build.sh is the only thing that catches ARM64 test-file problems.
Both were true when written and neither is now, so it is updated: the script is
the only *local* gate, CI covers it at run time, and the macOS ARM64 runner's
three processors are called out so the next hardcoded thread count is caught
before it is pushed.

Verified: gofmt, go vet, go build, go test -short ./..., go test -race -short
./internal/fit/..., check-cross-build.sh (six targets), golangci-lint
--new-from-merge-base=origin/main (0 issues), the renderer suite under
qemu-aarch64, and amd64 at GOMAXPROCS 1, 3 and 4.

+++ internal/fit/renderer/polish_dirty_cost.go
@@ -160,7 +160,10 @@ func (s *polishDirtySession) collectCircleSpans(circle fit.Circle, dirty *dirtyS
 		return
 	}
 
-	radiusSquared := circle.R * circle.R
+	// Rounded at the assignment, not only where it is used: this is a product,
+	// so leaving it unrounded lets arm64 fuse it straight into the subtraction
+	// below as FNMSUBD and decide a span boundary differently than amd64.
+	radiusSquared := float64(circle.R * circle.R)
 
 	fixed, useFixed := fixedCircleQ16{}, false
 	if !s.forceFloatGeometry && !s.forceFloat32Geometry {
@@ -182,7 +185,7 @@ func (s *polishDirtySession) collectCircleSpans(circle fit.Circle, dirty *dirtyS
 		center := float32(circle.X)
 		centerY := float32(circle.Y)
 		radius := float32(circle.R)
-		radiusSquared32 := radius * radius
+		radiusSquared32 := float32(radius * radius)
 
 		for y := minY; y < maxY; y++ {
 			dy := float32(y) - centerY
@@ -271,7 +274,10 @@ func (s *polishDirtySession) compositeCircleDirtyRows(circle fit.Circle, dirty *
 	// per dirty sub-span, so a single row could rebuild the block several times.
 	blend := newSpanBlend(circle.CR, circle.CG, circle.CB, circle.Opacity)
 
-	radiusSquared := circle.R * circle.R
+	// Rounded at the assignment, not only where it is used: this is a product,
+	// so leaving it unrounded lets arm64 fuse it straight into the subtraction
+	// below as FNMSUBD and decide a span boundary differently than amd64.
+	radiusSquared := float64(circle.R * circle.R)
 
 	fixed, useFixed := fixedCircleQ16{}, false
 	if !s.forceFloatGeometry && !s.forceFloat32Geometry {
@@ -293,7 +299,7 @@ func (s *polishDirtySession) compositeCircleDirtyRows(circle fit.Circle, dirty *
 		center := float32(circle.X)
 		centerY := float32(circle.Y)
 		radius := float32(circle.R)
-		radiusSquared32 := radius * radius
+		radiusSquared32 := float32(radius * radius)
 
 		for y := minY; y < maxY; y++ {
 			dy := float32(y) - centerY
